+++ test/checksum.test.ts
@@ -72,7 +72,7 @@ describe('Checksum', () => {
 		).toBe(1)
 	})
 
-	it('Always accept plugin when name is not provided', async () => {
+	it('Deduplicate global hook on use', async () => {
 		const cookie = (options?: Record<string, unknown>) =>
 			new Elysia({
 				seed: options
@@ -90,6 +90,54 @@ describe('Checksum', () => {
 
 		expect(
 			Math.abs(a.hooks.transform!.length - b.hooks.transform!.length)
+		).toBe(0)
+	})
+
+	it("Don't filter inline hook", async () => {
+		const cookie = (options?: Record<string, unknown>) =>
+			new Elysia({
+				seed: options
+			}).onTransform(() => {})
+
+		const group = new Elysia().use(cookie()).get('/a', () => 'Hi', {
+			transform() {}
+		})
+
+		const app = new Elysia()
+			.use(cookie())
+			.use(group)
+			.get('/cookie', () => 'Hi')
+
+		// @ts-ignore
+		const [a, b] = app.routes
+
+		expect(
+			Math.abs(a.hooks.transform!.length - b.hooks.transform!.length)
 		).toBe(1)
+	})
+
+	it('Merge global hook', async () => {
+		let count = 0
+
+		const cookie = (options?: Record<string, unknown>) =>
+			new Elysia({
+				seed: options
+			}).onTransform(() => {})
+
+		const group = new Elysia()
+			.use(cookie())
+			.onTransform(() => {
+				count++
+			})
+			.get('/a', () => 'Hi')
+
+		const app = new Elysia()
+			.use(cookie())
+			.use(group)
+			.get('/cookie', () => 'Hi')
+
+		await Promise.all(['/a', '/cookie'].map((x) => app.handle(req(x))))
+
+		expect(count).toBe(2)
 	})
 })